now directional and spot light produce directional shadow

+++ src/game/config.hpp
@@ -2,14 +2,12 @@
 #include <glm/vec2.hpp>
 
 struct Config {
-
 	// --- RESOLUTIONS ---
 	glm::ivec2 windowRes;
 	glm::ivec2 renderRes;
-	int shadowRes;
+	uint16_t shadowRes;
 
 	// --- GRAPHICS ---
 	float renderDistance;
 	bool vsync;
-
 };
